% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_plotarium}
\alias{run_plotarium}
\title{Launch Plotarium App}
\usage{
run_plotarium(launch.browser = TRUE, background = TRUE)
}
\arguments{
\item{launch.browser}{Logical; if TRUE, the app will open in the default web browser.}

\item{background}{Logical; if TRUE, the app will run in the background using the processx package.}
}
\value{
None; the function launches the Shiny app.
}
\description{
This function launches the Plotarium Shiny application. It can be run in the background or foreground.
}
